# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Detección de Plagas",
        "description": "Documentación de la API para el sistema de detección de plagas usando Laravel y YOLOv5",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/notificaciones": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Obtener las últimas 10 notificaciones del usuario autenticado",
                "operationId": "4a5488228077528a278bea003da9a3cc",
                "responses": {
                    "200": {
                        "description": "Lista de notificaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notificaciones": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "mensaje": {
                                                        "type": "string",
                                                        "example": "Tienes una nueva alerta"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-19T20:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-19T20:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Crear una nueva notificación",
                "operationId": "d2dd16a0c0f439306ecfa7ec0312972f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mensaje"
                                ],
                                "properties": {
                                    "mensaje": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Tienes una nueva alerta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notificación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "notificacion": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "mensaje": {
                                                    "type": "string",
                                                    "example": "Tienes una nueva alerta"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-19T20:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-19T20:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "cedula",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirección al login con mensaje de éxito"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener token Sanctum",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirección a /captura si las credenciales son válidas"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión y eliminar tokens Sanctum",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "302": {
                        "description": "Redirección a la página de inicio"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/captura": {
            "post": {
                "tags": [
                    "Captura"
                ],
                "summary": "Guardar imagen base64 y detectar plagas",
                "operationId": "9e7e4282137529451f8b04d7442aed17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imagen"
                                ],
                                "properties": {
                                    "imagen": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vista con imagen procesada y resultados de detección"
                    },
                    "400": {
                        "description": "No se recibió ninguna imagen"
                    },
                    "500": {
                        "description": "Error al comunicarse con el detector de plagas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/detecciones": {
            "post": {
                "tags": [
                    "Detecciones"
                ],
                "summary": "Registrar una detección de plaga",
                "operationId": "c941fcd02ee2d40c7e337af308c03987",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plaga",
                                    "ubicacion",
                                    "hora"
                                ],
                                "properties": {
                                    "plaga": {
                                        "type": "string",
                                        "example": "Mosca blanca"
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "example": "Invernadero 3, sector norte"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-19T14:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detección registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Detectada y guardada"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/plagas/detectar": {
            "post": {
                "tags": [
                    "Plagas"
                ],
                "summary": "Detectar plagas en una imagen base64 usando YOLOv5",
                "operationId": "7db10e707ecf8fc0e91891743f955bda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imagen"
                                ],
                                "properties": {
                                    "imagen": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vista con imagen procesada y resultados de detección"
                    },
                    "400": {
                        "description": "Formato de imagen inválido o error en la detección"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Notificaciones",
            "description": "Notificaciones"
        },
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Captura",
            "description": "Captura"
        },
        {
            "name": "Detecciones",
            "description": "Detecciones"
        },
        {
            "name": "Plagas",
            "description": "Plagas"
        }
    ]
}